Create IIP-User-Stat-Report.md
Draft Proposal of improved user report tools. I think it may be possible to split this into two IIP's, but on similar topics:
1. Improved and unified reporting of values in additional currency values and improved/better integrated charting and comparison tools.
2. Annual report production.
Bearing in mind the informal, community-sided nature of the IIP's, I don't know if it is worth discussing in what ways we would want the features to be implemented - i.e a separate 'performance' page on the top bar, or pdf sent via email, or whatever, but we can.